Migrated UL20 measure TCs
